Rename new_user_setups.py to setups.py and remove missed commented lines.
